Add type signatures for aggregation logic
This is the first real step towards #502.

Factoring this out because deciding the type signatures was non-trivial,
and the work was paired with @wchargin.

This also moves the EdgeType and NodeType types into core/graph, as the
aggregation logic should really live in core/attribution and this
removes a layer of indirection. I've added a TODO to formalize the
contract for edge and node type, and linked to #465.

Test plan: `yarn test`