Add a React context for resolving static assets
Summary:
This will enable components to obtain the path to a static asset, even
when the app is not hosted at the root of a server, as outlined in #643.
This module will be used for simple assets (images, etc.) and API data
(fetches from `/api/**`) alike.

Test Plan:
Unit tests included; `yarn test` suffices.

wchargin-branch: assets-context